Fixed Windows compatibility for WebDAV server. The server now works with Windows XP (Service Pack 2). Our XML reply for the PROPFIND request was not 100% correct (although it was working with other clients which are less picky/strict). It took some time to figure out this particular fix (like 1/2 day of painful packet sniffing and testing). Please let me or Amos know if you're experiencing problems. It should be working quite well now, I've tested with Win2K (no updates), WinXP+SP2, CaDAVer, Konqueror and Finder (Mac OSX). Some small issues remain: I think Win2K needs at least IE 2600 series in order to function properly (renaming doesn't work at some locations). Mac OSX: a strange permission problem was discovered during the evening (will probably fix later).
*** IMPORTANT ***

When mounting in WinXP: __REMEMBER__ to use the # (hash)
sign as the last character in the URL. If omitted, XP
thinks that you're trying to connect to a SAMBA share and
thus all the URLs (and logic behind the scenes) will be 
messed up: "\\horrible\protocol" instead of "http://webdav/rocks".
 
Mounting instructions:

1) Fire up IE
2) Select "Open" from the "File" menu
3) Type in http://webdav.yoursite.example.com/#
4) Check "Open as Web Folder"
5) Click "OK"
6) If everything goes well, the thing will be
   properly mounted and you can access it as as
   a network drive (look in your Networkhood).

Known Windows issues (in particular XP + SP2):
"net use * http://webdav.example.com/" 
will *NOT* work. This is a Windows problem, it 
will simply say "System error 67 has occured" without
even sending a single packet to the server. However, 
if you outsmart the sucker and try mounting like this:
"net use * http://webdav.example.com/whatever/"
..where "whatever" is the name of an existing folder that
you wish to reach, it will work. BUT! ->  The "net use" part
of Windows apparently doesn't make 100% use of the new WebDAV
Mini Redirector subsystem which came with SP2 for XP. There
is no support for the basic HTTP authentication scheme; which
is a pretty annoying issue since it simply renders the "net
use" command "net useless" when it comes to non-IIS web shares.
(In IIS you can turn off basic authentication and use digest or
 something else instead.)
 




git-svn-id: file:///home/patrick.allaert/svn-git/ezp-repo/ezpublish/trunk@10087 a01eee8c-daba-0310-acae-fa49f3023285