Fix #007591: Automatic virtual host mode detection fails
Un deprecate ForceVirtualHost to allow it top be used in case we are missing test data for some setup that still needs it.
But in theory virtual host detection should be automatic now.
Also cleanup unused ServerParameters and make sure we check that phpSelf contains index.php before we try to do any VH / NVH detection.