Fix non working search dialog caused by jQuery 1.5 update
Short: Form element names that end in [] where encoded by jQuery.param(), hence did not work anymore.
Workaround is to pass a string to jQury's ajax functions instead of array / object to make sure the param() function is not used.

Also a small fix for bookmarks to make sure count on list of objects are only used if total count is larger then 0.