Updated comment format for method
Following on from my first attempts to add documentation to the methods
of classes in the kernel classes.

I have now followed the request in comments to set it to null.

When I code in php I do personally use null, however I was curious why
eclipse in php mode by default gave me NULL.

If we goto the php documentation.

http://www.php.net/manual/en/language.types.null.php

We see clearly that the reason the ide gave me NULL is becuase that is
what the real php syntax is all about.

I learnt something new today.