Make use of the already implemented type hinting
The array type hinting in the constructor already makes sure that the parameter *must* be an array. So it is not event possible to use the null value as in `$sys = new eZSys( null );` - this would throw a catchable fatal error.

Now sure to have an array at any time, prettify the code a bit :).