Implemented #18839: Apply Script Monitor patch for subtree removal
The Script Monitor extension contains a subtree removal patch which ensures that removals are deferred to cron when the amount of sub nodes is above
the threshold given in content.ini [RemoveSettings] MaxNodesRemoveSubtree. This patch should be committed to master.

If the Script Monitor extension is not active, the previous functionality is used: Either remove immediately or refuse removal, depending on the
MaxNodesRemoveSubtree setting.