Removed the line of ezpDatabaseHelper that emptied the database name in the db params.
I remember encountering a complete blocker on this line when first testing the cluster
with postgres. I have investigated, but still can't see why we need to do this.

One problem with postgres is that unlike MySQL, connect + db choice = 1 action, not 2.
If pg_connect isn't provided with the dbname, it searches for a db named like the user.

Let me know if you know more :)