Fixed issue #19668: Problem with viewcache expiry
in rare cases, during expiry of caches on DFS, the update of the ezdfsfile.mtime table will fail and store a value of zero.
according to mysql documentation if the ABS(x) throws an error, it returns zero, the query does not fail.

as a consequence mtime=0 will corrupt the row and the cache cannot be regenerated.