array_unique() on 2-dimensional array produces PHP notices
When accessing /content/history/<node_id> with PHP Notices display enabled, an error Array to string conversion is shown for each displayed version.

This is because of the usage of array_unique() which is not intended to be used on multi-dimensional array (see http://php.net/array_unique in the "Notes" section).

This fix uses the approach proposed at http://stackoverflow.com/a/946300/284325 to enable the wanted behaviour for the versionArray