Update kernel/classes/clusterfilehandlers/ezfsfilehandler.php
Changed purge function to use a depth-first search instead of a breadth-first search when scanning the file systems for cache files to delete. This will make the code use much less memory and run faster if there are a large number of files. Before it tried to store all files in an array and that caused php to crash with a memory allocation error and the cache clearing job to fail.