EZP-20813: Added NFS mount checks in dfscleanup.php
EZP-20813: dfscleanup.php will now abort on DB/NFS error

The script now checks if NFS is available:
- on startup
- after a file has been found as non-existent, before erasing the DB record

This is done by checking existence + type on <NfsMount>/<vardir>.

For the DB, When an unexpected error occurs within eZDFSFileHandlerMySQLiBackend,
an error was logged and false was returned. Since this method is
used by exists(), you get the same error if a file doesn't exist
or if the DB is just dead.

In dfscleanup.php, this led to the script being able to wipe the NFS
folder because of a database error on exist().

When such an error occurs, an exception will be raised, that will be
caught by the script, and the script will be aborted.

EZP-20813: review remarks