EZP-21324: Fixed missing references to copied image attributes
- eZImageType::postStore() adds references from ezimage attributes
- An update script (update/common/5.3/restoreimageslinks.php will add missing
  relations for existing data

Also changed removeAlias code. Should now check correctly:
- only remove ezimagealias entries when they're not used anymore
  (handles duplicates if any)
- only removed files when not used anymore.

This means that we accept that when a copy of a content with an
image is made, we accept that archived versions in the copy will
reference the same file than archive versions in the copy source.

EZP-21324: changed removeAlias() prototype

Doesn't accept an eZContentObjectAttribute argument anymore.
The method is always invoked on the instance it must be modified.

EZP-21324: Fixed postgres error

fetchListByFilePath now uses asObject=false, since it doesn't
really make sense to return incomplete objects (due to the duplicates
inserted in ezimagefile)

Fixed removeAllAliases()

We just don't use it anymore. removeAliases() does the job.
The only difference is that removeAliases() will set the
attribute's XML to an empty value, while removeAllAliases()
won't, meaning that the change will create a light overhead
on the database.

EZP-21324: Prevented duplicates in ezimagefile

`eZImageType::postStore()` won't add duplicates to ezimagefile anymore.