Fix "Invalid package" issue of the installation wizard
When installing legacy eZ Publish with the installation wizard, the package installation always fail with the "Invalid package" error. 

Because in line 1163, $entry->getPath() returns an entry path with leading '/', but eZPackage::definitionFilename() is "package.xml", so "package.xml" will never be extracted.

The "Invalid package" issue is also related to another class in ezc, which I will post another fix.