EZP-22321: ezjscore cache id not cleared with dfs
Fixed eZDFSFileHandler mysqli implementation of delete.

The interface states that directories are deleted, but the implementation only covered single files. It has been updated to delete a single file if there is one, and a set of files if there is none. Set of files can be deleted using name_trunk, if it applies, or using a wildcard if it doesn't.

A new name_trunk identifier has been added for ezjscore.