Fix EZP-22466: adv. obj. relation list edits not saved
Fixes the condition at the top of eZObjectRelationList::validateObjetAttributeHTTPInput().
It should now only skip validation if the attribute isn't required AND doesn't contain any relation.

The issue was a regression from https://jira.ez.no/browse/EZP-18790. The bug fixed in this issue
was tested, and no regression was found.