Fix EZP-2881: AcceptPathInfo Test bug
Removing this test as it has several issues and is not reliable,
it shows up almopst always causing frustation for new users and
makes it difficult to do accaptance testing of setup wizard.

As for the 2 reasons why it exists:
1: Since it was created fast cgi support was added.
2: Additionally, neither eZ Publish legacy or Symfony uses
PATH_INFO under Apache which is the reason to enable
AccptPathInfo in the first place.

So in summary: This test is obsolete.