[Travis] Change test matrix to test only once per php version and db
Moved the timezones to be random per build instead of baing in the
matrix, also reduce number of bulds so we cover each php version
once and actoss them a mix of mysql and posgres.

Motivation is to make travis faster, easier to maintain the travis.yml
and make space for 5.6 and hhvm.