Fixed EZP-22974: When I create a new Image I get Error: eZDir::recursiveDelete
The error occurs when clearing cache that does not exist (has not been generated). I have not removed the call to clearing cache, since I don't know for sure if there are cases where cache does exist. Instead the cache clearing code is improved so that it doesn't attempt to delete non-existing directories.

Now it silently ignores such calls. I don't see any need to log them - it is perfectly normal that this happens.

https://jira.ez.no/browse/EZP-22974