No longer clear the image-alias cache by default
It is quite expensive to clear the image-alias cache on a site, doing
so requires re-creating all image variations for new page loads.
Clearing of this cache is primarily meant for kernel developers testing
out new variation techniques, so there is no need for it to be cleared
by default.
Now ezcache.php is changed to only clear cache entries that are marked
as `default`, image-alias has been changed to be non-default.

A new `--clear-non-default` option is available to clear only those
that are non-default. It is also still possible to clear it specifically
using the `--clear-id` option.