EZP-26680 INI Cache files are generated with wrong filepaths
When publicAPI calls are executed via Symfony controller and for example a user is created via UserService, this Signal is triggered into legacy System cache management.

The eZINI::instance() is called inside user datatype which checks site.ini cache file. In the cache file are relative paths to override ini files which do not work because application is in Symfony context (and not cwd in legacy).

The override file is not found via file_exists() and the new generated cachefile is w/o override files cached, which "destroys" legacy application.