[lib/version] Configure master branch back to dev
Version is set to 2019.99-alpha1 as there isn't plans atm for what next
feature release will be if any, if there is needs for another feature
release community would need to drive that.

<ad>
With eZ Platform 2.5LTS being out and bringing more or less feature parity
with eZ Publish it's a good time to evaluate it again. With legacy bridge
support it's also a good version to migrate to if you have legacy needs.

Or if you are a developer then eZ Platform 3 comming this year might be of
interest for new projects, as one of it's core aims is to greatly improve
developer experience, closer integrate the CMS with Symfomny & lots more.
</ad>

Thanks for all the fish, I mean legacy!