Added support for PSR loggers in eZDebug
By setting a global variable all internal logging in eZDebug
will be turned off and instead it will divert the messages to
the configured logger channels.

Set `$GLOBALS['ezpDebug']['loggerFactory']` to a callable which
accepts a parameter for the level type and should return a the
logger channel instance, this can be a single instance for all
levels or one per level. eZDebug will then use this to log the
message using the PSR interface.
To disable a log level the callable should return null.

A simple example using Monolog that will log to a single file.

```php
$GLOBALS['ezpDebug']['loggerFactory'] = function ($level) {
    static $log;
    if ($log === null) {
        $log = new \Monolog\Logger('app');
        $log->pushHandler(new \Monolog\Handler\StreamHandler('app.log'));
    }
    return $log;
};
```