Avoid throwing errors when clearing cache dirs fails
For high traffic sites the cache folder may be written to while
the current process is busy removing cache files. This results
in the inability to remove the cache folder which was presumed
to be empty.
We catch this case and simple ignore the error as the old cache
files have already been successfully removed.